split up wayland.rs to make it easier to work on
